(all) silence compile warnings on Windows

+++ mfscommon/crc.c
@@ -148,7 +148,7 @@ uint32_t mycrc32(uint32_t crc,const uint8_t *block,uint32_t leng) {
 #endif
 	CRC_REORDER;
 #ifdef WIN32
-	while (leng && ((unsigned long long int)block & 7)) {
+	while (leng && ((uintptr_t)block & 7)) {
 #else
 	while (leng && ((unsigned long int)block & 7)) {
 #endif
